requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.1.7
django-allauth==65.7.0
django-crispy-forms==2.3
django-filter==25.1
sqlparse==0.5.3
tzdata==2025.2